rename: #8 - management 네이밍 변경
